requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
h11==0.16.0
idna==3.11
joblib==1.5.3
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
python-dotenv==1.0.1
sentence-transformers
twilio